BVE5: Minor tweaks
[ci skip]
