+++ source/OpenBVE/OpenBve.csproj
@@ -162,6 +162,7 @@
     <Compile Include="Graphics\Renderer\Structures.cs" />
     <Compile Include="Graphics\Renderer\Objects.cs" />
     <Compile Include="Parsers\Object\Generic\WavefrontObjParser.cs" />
+    <Compile Include="Parsers\Route\BVE5\Bve5ScenarioParser.ApplyRouteData.cs" />
     <Compile Include="Parsers\Route\BVE5\Bve5ScenarioParser.Commands.cs" />
     <Compile Include="Parsers\Route\BVE5\Bve5ScenarioParser.Components.cs" />
     <Compile Include="Parsers\Route\BVE5\Bve5ScenarioParser.cs" />
